Prototipo de clase tick

+++ Modelo/src/modelo/Acompanamiento.java
@@ -13,7 +13,7 @@ public class Acompanamiento extends Producto{
     
     public String tipo;     //salado o dulce
 
-    public Acompanamiento(String nombre, int precio, double tiempoProduccion, String tipo) {
+    public Acompanamiento(String nombre, int precio, int tiempoProduccion, String tipo) {
         super(nombre, precio, tiempoProduccion);
         this.tipo = tipo;
     }
